+++ src/main/java/com/AniCare/demo/Dto/medical/VetReviewSetDto.java
@@ -5,10 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@AllArgsConstructor@NoArgsConstructor@Setter@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Setter
+@Getter
 public class VetReviewSetDto {
 
     private Long clinicId;  // 평가할 진료 id
-    private float rating;   // 부여할 별점
+    private double rating;   // 부여할 별점
 
 }
